♻️ Refactor: districtService 내부 커스텀훅 분리
